chatter_to_chapter/requirements.txt
google-adk[all]
google-genai
deprecated
Pillow
